+++ apps/opentelemetry_experimental/test/otel_metrics_SUITE.erl
@@ -97,7 +97,8 @@ init_per_testcase(provider_test, Config) ->
     application:load(opentelemetry_experimental),
     ok = application:set_env(opentelemetry_experimental, readers,
                              [#{module => otel_metric_reader,
-                                config => #{exporter => {otel_metric_exporter_pid, self()}}}]),
+                                config => #{exporter => {otel_metric_exporter_pid, self()},
+                                            default_temporality_mapping => default_temporality_mapping()}}]),
     ok = application:set_env(opentelemetry_experimental, views,
                              [#{selector => #{instrument_name => a_counter},
                                 aggregation_module => otel_aggregation_sum},
@@ -117,9 +118,11 @@ init_per_testcase(multiple_readers, Config) ->
     application:load(opentelemetry_experimental),
     ok = application:set_env(opentelemetry_experimental, readers,
                              [#{module => otel_metric_reader,
-                                config => #{exporter => {otel_metric_exporter_pid, self()}}},
+                                config => #{exporter => {otel_metric_exporter_pid, self()},
+                                            default_temporality_mapping => default_temporality_mapping()}},
                               #{module => otel_metric_reader,
                                 config => #{exporter => {otel_metric_exporter_pid, self()},
+                                            default_temporality_mapping => default_temporality_mapping(),
                                             default_aggregation_mapping =>
                                                 #{?KIND_COUNTER => otel_aggregation_drop}}}]),
 
@@ -132,13 +135,14 @@ init_per_testcase(delta_counter, Config) ->
     %% delta is the default for a counter with sum aggregation
     %% so no need to set any temporality mapping in the reader
     ok = application:set_env(opentelemetry_experimental, readers, [#{module => otel_metric_reader,
-                                                                     config => #{exporter => {otel_metric_exporter_pid, self()}}}]),
+                                                                     config => #{exporter => {otel_metric_exporter_pid, self()},
+                                                                                 default_temporality_mapping => default_temporality_mapping()}}]),
 
     {ok, _} = application:ensure_all_started(opentelemetry_experimental),
 
     Config;
 init_per_testcase(cumulative_counter, Config) ->
-    CumulativeCounterTemporality = #{?KIND_COUNTER =>?TEMPORALITY_CUMULATIVE},
+    CumulativeCounterTemporality = maps:put(?KIND_COUNTER, ?TEMPORALITY_CUMULATIVE, default_temporality_mapping()),
     application:load(opentelemetry_experimental),
     ok = application:set_env(opentelemetry_experimental, readers, [#{module => otel_metric_reader,
                                                                      config => #{exporter => {otel_metric_exporter_pid, self()},
@@ -149,7 +153,7 @@ init_per_testcase(cumulative_counter, Config) ->
 
     Config;
 init_per_testcase(delta_explicit_histograms, Config) ->
-    DeltaHistogramTemporality = #{?KIND_HISTOGRAM =>?TEMPORALITY_DELTA},
+    DeltaHistogramTemporality = maps:put(?KIND_HISTOGRAM, ?TEMPORALITY_DELTA, default_temporality_mapping()),
     application:load(opentelemetry_experimental),
     ok = application:set_env(opentelemetry_experimental, readers, [#{module => otel_metric_reader,
                                                                      config => #{exporter => {otel_metric_exporter_pid, self()},
@@ -160,7 +164,7 @@ init_per_testcase(delta_explicit_histograms, Config) ->
 
     Config;
 init_per_testcase(delta_observable_counter, Config) ->
-    DeltaObservableCounterTemporality = #{?KIND_OBSERVABLE_COUNTER =>?TEMPORALITY_DELTA},
+    DeltaObservableCounterTemporality = maps:put(?KIND_OBSERVABLE_COUNTER, ?TEMPORALITY_DELTA, default_temporality_mapping()),
     application:load(opentelemetry_experimental),
     ok = application:set_env(opentelemetry_experimental, readers, [#{module => otel_metric_reader,
                                                                      config => #{exporter => {otel_metric_exporter_pid, self()},
@@ -173,7 +177,8 @@ init_per_testcase(delta_observable_counter, Config) ->
 init_per_testcase(_, Config) ->
     application:load(opentelemetry_experimental),
     ok = application:set_env(opentelemetry_experimental, readers, [#{module => otel_metric_reader,
-                                                                     config => #{exporter => {otel_metric_exporter_pid, self()}}}]),
+                                                                     config => #{exporter => {otel_metric_exporter_pid, self()},
+                                                                                 default_temporality_mapping => default_temporality_mapping()}}]),
 
     {ok, _} = application:ensure_all_started(opentelemetry_experimental),
 
@@ -192,6 +197,16 @@ default_resource(_Config) ->
                  otel_attributes:map(otel_resource:attributes(Resource))),
 
     ok.
+
+default_temporality_mapping() ->
+    #{
+        ?KIND_COUNTER => ?TEMPORALITY_DELTA,
+        ?KIND_OBSERVABLE_COUNTER => ?TEMPORALITY_CUMULATIVE,
+        ?KIND_UPDOWN_COUNTER => ?TEMPORALITY_DELTA,
+        ?KIND_OBSERVABLE_UPDOWNCOUNTER => ?TEMPORALITY_CUMULATIVE,
+        ?KIND_HISTOGRAM => ?TEMPORALITY_DELTA,
+        ?KIND_OBSERVABLE_GAUGE => ?TEMPORALITY_CUMULATIVE
+    }.
 
 using_macros(_Config) ->
     DefaultMeter = otel_meter_default,